backend/requirements.txt
fastapi>=0.104.0,<0.110.0
uvicorn[standard]>=0.24.0,<0.30.0
motor>=3.3.2,<4.0.0
pymongo>=4.6.0,<5.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.10.1,<3.0.0
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0
langchain>=1.2.10,<2.0.0
langchain-community>=0.4.1,<1.0.0
langchain-core>=1.2.17,<2.0.0
langchain_openrouter>=0.1.0,<1.0.0
PyPDF2>=3.0.1,<4.0.0
python-docx>=0.8.11,<1.0.0
bcrypt>=4.1.1,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
httpx>=0.25.2,<1.0.0
